Tweaked & optimized bashmarks.sh
There's no shebang, so I'm not sure which shell you're shooting for; I'll assume bash. Hope this helps.

L50
No need for `sed`; Bash itself can do easily do this, via the parameter expansion's substitution feature (see `man bash`). If you want it to be global, change `/` to `//`.

L59
Not sure I follow the reasoning for what was there prior to my suggestion, but what I propose looks to have the very same effect, without the redundancy, but you'll probably want to test beforehand. The same approach can be applied elsewhere, such as L73.

L88-96
A case statement simplifies this block, making it much easier to work with, IMO. I'd personally also use a heredoc (`<<-` for readability) instead of all of those `echo` calls, to further simplify. I also added `-?` because, while it's not as common, it is sometimes used out in the wild.

L114
If you're looking for the path, sans final element, then you could use either `dirname`, or, preferably, make use of bash parameter expansion in a while loop, as demonstrated, to avoid unnecessarily opening up more processes. The approach I've used here is one of a pure-shell style. You needn't worry about performance issues unless you're dealing with many thousands of ^DIR environment variables, which I imagine to be unlikely, and perhaps impractical. You could further optimize this by writing this block all in `awk` to avoid using `sort`, but then you rely on that, so I personally wouldn't go that route, if portability is very important to you, since the former is part of coreutils, but I believe the latter isn't.

L134
Cleaned up function `_comp()"`, removing unneeded code. The `local` built-in can assign and declare simultaneously, as with `declare`.

L141
This function gets processed and loaded even if you don't have or are not on ZSH; I think this should be something you test for, to keep things clean. It's been a while since I've used ZSH, but if I remember correctly, it also has a SHELL type environment variable you can test for.

L101
`1` is "not a valid identifier" for a function. This should be reconsidered, unless it's just a typo.